api/pyproject.toml
[project]
name = "summarize"
version = "0.1.1"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "peewee (>=3.17.9,<4.0.0)",
    "flask-cors (>=5.0.0,<6.0.0)",
    "yt-dlp (>=2025.1.26,<2026.0.0)",
    "structlog>=25.1.0",
    "openai (>=1.58.1,<2.0.0)",
    "python-dotenv>=1.2.1",
]

[tool.uv]
package = false

[tool.basedpyright]
typeCheckingMode = "off"
